requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
pyotp
qrcode
pillow
tensorflow
numpy
requests
